refact(CGeneralizedCliffordsBSimulationStrategy): common functions
There were several functionalities in the Generealized Cliffords B Simulation Strategies
and those functionalities were moved to a separate common file: samplingHelperFunctions.
This contains the DFE calls of the permanent calculators which results in better maintenance
for this feature.
