+++ piquassoboost/sampling/simulation_strategies/source/CGeneralizedCliffordsBSimulationStrategy.h
@@ -42,7 +42,7 @@ class CGeneralizedCliffordsBSimulationStrategy {
 protected:
 
     /// The number of photons
-    int64_t number_of_input_photons;
+    //int64_t number_of_input_photons;
     /// The matrix describing the interferometer
     matrix interferometer_matrix;
     /// The input state entering the interferometer
@@ -117,17 +117,6 @@ std::vector<PicState_int64> simulate( PicState_int64 &input_state_in, int sample
 protected:
 
 /**
-@brief Call to randomly increase the current input state by a single photon
-*/
-void update_current_input();
-
-
-/**
-@brief Call to calculate new layer of probabilities from which an intermediate (or final) output state is sampled
-*/
-void compute_pmf(PicState_int64& sample);
-
-/**
 @brief Call to recursively add substates to the hashmap of labeled states.
 */
 //void append_substate_to_labeled_states( PicState_int64& iter_value);
@@ -137,7 +126,7 @@ void compute_pmf(PicState_int64& sample);
 @brief Call to calculate and fill the output states for the individual shots.
 @param sample The current sample state represented by a PicState_int64 class
 */
-void fill_r_sample( PicState_int64& sample );
+void fill_r_sample( PicState_int64& sample, int64_t number_of_input_photons );
 
 
 /**
@@ -149,11 +138,6 @@ void fill_r_sample( PicState_int64& sample );
 
 
 
-/**
-@brief Call to pick a new sample from the possible output states according to the calculated probability distribution stored in pmfs.
-@param sample The current sample represanted by a PicState_int64 class that would be replaced by the new sample.
-*/
-void sample_from_pmf( PicState_int64& sample );
 
 
 }; //CGeneralizedCliffordsBSimulationStrategy
